backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
langchain
langchain-core
langchain-groq
groq
pydantic
fastapi[all]
python-multipart
pytesseract
Pillow
pypdf
pandas
openpyxl
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
